+++ src/popup/content/App.tsx
@@ -1,13 +1,12 @@
 import { defineComponent, ref, onMounted, provide, Transition } from "vue";
+import { RefreshOutline } from "@vicons/ionicons5";
 import browser from "webextension-polyfill";
 import Head from "./components/Head";
 import Footer from "./components/Footer";
-import SettingPanel from "./components/SettingPanel";
-import {
-    ChevronBackOutline,
-} from "@vicons/ionicons5";
-import { useConfig, useDegBtn } from "./use";
-import { initRollConfig, updateRollConfig } from "./utils";
+import GridPanel from './components/GridPanel';
+import Info from './components/Info';
+import { useConfig } from "./use";
+import { initRollConfig, updateRollConfig, reloadPage } from "./utils";
 import { clone } from "../../util";
 import { ActionType } from "../../types/type.d";
 
@@ -28,14 +27,14 @@ export default defineComponent({
         // current website config
         const rollConfig = useConfig();
 
-        // buttons
-        const degBtns = useDegBtn();
-
         const update = updateRollConfig.bind(null, rollConfig);
         provide("rollConfig", rollConfig);
         provide("update", update);
         provide("onOpenSetting", onOpenSetting);
 
+        const onReload = () => {
+            reloadPage(rollConfig.tabId);
+        }
         /**
          * 当打开时就获取当前网站的视频信息
          * 添加样式
@@ -57,15 +56,24 @@ export default defineComponent({
             )
 
             chrome.runtime.onMessage.addListener((a, b, c) => {
-                const { type, rollConfig: config } = a;
-                if (type === ActionType.UPDATE_STORAGE) {
-                    Object.keys(config).forEach((key) => {
-                        if (key in rollConfig) {
-                            rollConfig[key] = config[key];
-                        }
-                    });
+                const { type, rollConfig: config, text } = a;
+                switch(type) {
+                    case ActionType.UPDATE_STORAGE:
+                        console.log(config, '---config');
+                        Object.keys(config).forEach((key) => {
+                            if (key in rollConfig) {
+                                rollConfig[key] = config[key];
+                            }
+                        });
+                        break;
+                    case ActionType.UPDATE_BADGE:
+                        rollConfig.videoNumber = Number(text);
+                        break;
+                    default:
+                        break;
                 }
-                c("update_storage");
+
+                c("update");
             });
         });
 
@@ -73,34 +81,22 @@ export default defineComponent({
             <div>
                 <Head isShow={isShow.value}></Head>
                 <main>
-                    <div>
-                        <div class="video-roll-content">
-                            <div class="video-roll-website">
-                                <span>{rollConfig.name}</span>
-                            </div>
-                            <div class="video-roll-rotate-control">
-                                {degBtns.value.map((item) => (
-                                    <div
-                                        class={`rotate-${item.type}-${item.iconDeg} rotate-btn`}
-                                        key={item.type}
-                                        onClick={() => update("deg", item.deg)}
-                                    >
-                                        <ChevronBackOutline />
-                                    </div>
-                                ))}
-
-                                <div class="rotate-deg-text">
-                                    {rollConfig.deg}
-                                </div>
-                            </div>
+                    <div class="video-roll-headBar">
+                        <Info></Info>
+                        <div title="reload page" class="reload-btn" onClick={onReload}>
+                            <RefreshOutline class="video-roll-icon"></RefreshOutline>
                         </div>
-                        <Footer></Footer>
                     </div>
-                    <Transition name="van-fade">
+                   
+                    <div class="video-roll-content">
+                        <GridPanel></GridPanel>
+                    </div>
+                    <Footer></Footer>
+                    {/* <Transition name="van-fade">
                         <div class="video-roll-setting" v-show={isShow.value}>
                             <SettingPanel />
                         </div>
-                    </Transition>
+                    </Transition> */}
                 </main>
             </div>
         );