+++ src/popup/content/components/Focus/index.tsx
@@ -3,8 +3,9 @@
  * @Author: Gouxinyu
  * @Date: 2022-09-19 22:53:23
  */
-
+<ion-icon name="eye-outline"></ion-icon>
 import { defineComponent, inject } from "vue";
+import { EyeOutline, EyeOffOutline } from "@vicons/ionicons5";
 import type { IRollConfig } from "../../../../types/type.d";
 import "./index.less";
 
@@ -14,24 +15,19 @@ export default defineComponent({
         const update = inject("update") as Function;
         const rollConfig = inject("rollConfig") as IRollConfig;
 
-        const setFocus = (value: boolean) => {
-            rollConfig.focus.on = value;
+        const setFocus = () => {
+            rollConfig.focus.on = !rollConfig.focus.on;
             update("focus", rollConfig.focus);
         };
         return () => (
-            <div class="video-roll-focus">
-                <div class="focus-switch-box">
-                    <span class="focus-label">
-                        off
-                    </span>
-                    <van-switch
-                        class="focus-switch-btn"
-                        v-model={rollConfig.focus.on}
-                        size="22px"
-                        onUpdate:modelValue={setFocus}
-                    />
-                    <span class="focus-label">
-                        on
+            <div title='Focus mode' class={`video-roll-focus video-roll-item ${rollConfig.focus.on ? 'video-roll-on' : 'video-roll-off'}`} onClick={setFocus}>
+                <div class="video-roll-icon-box">
+                    <span class="video-roll-label">
+                        {
+                            rollConfig.focus.on ? <EyeOutline
+                            class="video-roll-icon"
+                        ></EyeOutline> : <EyeOffOutline class="video-roll-icon"></EyeOffOutline>
+                        }
                     </span>
                 </div>
             </div>