+++ src/popup/content/components/Pitch/index.tsx
@@ -5,6 +5,7 @@
  */
 
 import { defineComponent, inject } from "vue";
+import { PulseOutline } from "@vicons/ionicons5";
 import type { IRollConfig } from "../../../../types/type";
 import { getDefaultConfig } from '../../use';
 import "./index.less";
@@ -16,38 +17,42 @@ export default defineComponent({
         const rollConfig = inject("rollConfig") as IRollConfig;
 
         const setPitch = (value: number) => {
-            rollConfig.pitch = value;
+            rollConfig.pitch.value = value;
             update("pitch", rollConfig.pitch);
         };
 
-        const reset = () => {
-            setPitch(getDefaultConfig().pitch)
+        const setPitchOn = () => {
+            rollConfig.pitch.on = !rollConfig.pitch.on;
+            if (!rollConfig.pitch.on) rollConfig.pitch.value = 0;
+            update("pitch", rollConfig.pitch);
         };
 
         return () => (
             <>
-                <van-button class="video-roll-resetBtn" size="mini" icon="replay" type="primary" onClick={reset}>reset</van-button>
-                <div class="video-roll-pitch">
-                    <span class="zoom-label">low</span>
-                    <van-slider
-                        v-model={rollConfig.pitch}
-                        min={-1}
-                        max={1}
-                        step={0.01}
-                        bar-height="4px"
-                        onUpdate:modelValue={setPitch}
-                        v-slots={{
-                            button: () => (
-                                <div class="custom-button">
-                                    {rollConfig.pitch}
-                                </div>
-                            ),
-                        }}
-                    ></van-slider>
-                    <span class="zoom-label">high</span>
+                <div class="video-roll-long-box">
+                    <div class={`video-roll-switch ${rollConfig.pitch.on ? 'video-roll-switch-on':'video-roll-switch-off'}`} onClick={setPitchOn}>
+                        <PulseOutline class="video-roll-icon"></PulseOutline>
+                    </div>
+                    <div class="video-roll-pitch">
+                        <van-slider
+                            v-model={rollConfig.pitch.value}
+                            min={-1}
+                            max={1}
+                            step={0.01}
+                            bar-height="4px"
+                            disabled={!rollConfig.pitch.on}
+                            onUpdate:modelValue={setPitch}
+                            v-slots={{
+                                button: () => (
+                                    <div class="custom-button">
+                                        {rollConfig.pitch.value}
+                                    </div>
+                                ),
+                            }}
+                        ></van-slider>
+                    </div>
                 </div>
             </>
-
         );
     },
 });