+++ src/popup/content/use/useComponents.ts
@@ -4,39 +4,221 @@
  * @Date: 2022-09-11 11:12:50
  */
 import { shallowReactive } from 'vue';
+import {
+    defineComponent,
+    ref,
+    defineAsyncComponent,
+    resolveComponent,
+    h,
+} from "vue";
+
+interface IConfig {
+    type: string;
+    title?: string;
+    style?: Object;
+    children?: any[];
+}
+
+interface IComponentConfig extends IConfig {
+    type: 'component',
+    component: any;
+}
+
+interface IContainerConfig extends IConfig {
+    type: 'container',
+    title?: string,
+    col?: number;
+    showTitle?: boolean;
+    children?: IComponentConfig[] | IRowConfig[]
+}
+
+interface IRowConfig extends IConfig {
+    type: 'row';
+    children: IContainerConfig[]
+}
+
+interface ISwiperConfig extends IConfig {
+    type: 'swiper';
+    children: IRowConfig[]
+}
 
 function useComponents() {
-    const components = shallowReactive([
+    const components = shallowReactive<ISwiperConfig[]>([
         {
-            title: "Focus"
+            type: 'swiper',
+            children: [{
+                type: 'row',
+                style: {
+                    margin: '30px 0',
+                    height: '120px'
+                },
+                children: [
+                    {
+                        type: 'container',
+                        col: 16,
+                        title: "Rotate",
+                        showTitle: true,
+                        children: [{
+                            type: 'component',
+                            component: defineAsyncComponent(() => import("../components/Rotate"))
+                        }]
+                    },
+                    {
+                        type: "container",
+                        col: 8,
+                        title: 'Cache Site',
+                        showTitle: true,
+                        style: {
+                            flexDirection: "column",
+                            background: 'none',
+                            justifyContent: 'space-between',
+                        },
+                        children: [
+                            {
+                                type: 'row',
+                                style: {
+                                    marginBottom: '30px'
+                                },
+                                children: [
+                                    {
+                                        type: 'container',
+                                        title: 'Focus',
+                                        showTitle: true,
+                                        col: 24,
+                                        children: [
+                                            {
+                                                type: 'component',
+                                                component: defineAsyncComponent(() => import("../components/Focus"))
+                                            }
+                                        ]
+                                    },
+                                ]
+                            },
+                            {
+                                type: 'row',
+                                children: [
+                                    {
+                                        type: 'container',
+                                        title: 'Store',
+                                        showTitle: false,
+                                        col: 24,
+                                        children: [
+                                            {
+                                                type: 'component',
+                                                component: defineAsyncComponent(() => import("../components/Store"))
+                                            }
+                                        ]
+                                    }
+                                ]
+                            }
+                        ]
+                    }
+                ]
+            },
+            {
+                type: 'row',
+                style: {
+                    margin: '30px 0',
+                    height: '40px'
+                },
+                children: [
+                    {
+                        type: 'container',
+                        title: "Repostion",
+                        showTitle: true,
+                        col: 8,
+                        children: [{
+                            type: 'component',
+                            component: defineAsyncComponent(() => import("../components/Repostion"))
+                        }]
+                    },
+                    {
+                        type: 'container',
+                        col: 8,
+                        title: "Stretch",
+                        showTitle: true,
+                        children: [{
+                            type: 'component',
+                            component: defineAsyncComponent(() => import("../components/Stretch"))
+                        }]
+                    },
+                    {
+                        type: 'container',
+                        col: 8,
+                        title: "Flip",
+                        showTitle: true,
+                        children: [{
+                            type: 'component',
+                            component: defineAsyncComponent(() => import("../components/Flip"))
+                        }]
+                    },
+                ],
+            },
+            {
+                type: 'row',
+                style: {
+                    margin: '30px 0',
+                    height: '40px'
+                },
+                children: [
+                    {
+                        type: 'container',
+                        col: 24,
+                        title: "Zoom",
+                        showTitle: true,
+                        children: [{
+                            type: 'component',
+                            component: defineAsyncComponent(() => import("../components/Zoom"))
+                        }]
+                    },
+                ]
+            },
+            {
+                type: 'row',
+                style: {
+                    margin: '30px 0',
+                    height: '40px'
+                },
+                children: [
+                    {
+                        type: 'container',
+                        title: "Pitch",
+                        showTitle: true,
+                        col: 24,
+                        children: [{
+                            type: 'component',
+                            component: defineAsyncComponent(() => import("../components/Pitch"))
+                        }]
+                    },
+                ]
+            }]
         },
         {
-            title: "Pitch"
-        },
-        {
-            title: "Zoom"
-        },
-        {
-            title: "Move",
-        },
-        {
-            title: "Scale"
-        },
-        {
-            title: "Flip"
-        },
-        {
-            title: "Filter",
-        },
-        {
-            title: "Store"
-        },
-        {
-            title: "About"
-        },
+            type: 'swiper',
+            children: [
+            {
+                type: 'row',
+                style: {
+                    margin: '30px 0',
+                    height: '40px'
+                },
+                children: [
+                    {
+                        type: 'container',
+                        title: "Filter",
+                        showTitle: true,
+                        col: 8,
+                        children: [{
+                            type: 'component',
+                            component: defineAsyncComponent(() => import("../components/Filter"))
+                        }]
+                    }
+                ],
+            }]
+        }
     ]);
 
     return components;
 }
 
-export { useComponents }
+export { useComponents, ISwiperConfig, IRowConfig, IComponentConfig, IContainerConfig, IConfig }